requirements.txt
pyyaml
requests
pre-commit
flake8
flake8-bugbear
black
